Controllers Refactored

+++ src/main/java/com/freshlink/fishdto/FishCreateRequest.java
@@ -1,9 +1,11 @@
 package com.freshlink.fishdto;
 
+import java.math.BigDecimal;
+
 public record FishCreateRequest(
 		String name,
 		String fishTypeName,
-        double pricePerKg,
+		BigDecimal  pricePerKg,
         double availableKg
 		) {
