+++ src/main/java/com/freshlink/fishdto/FishResponse.java
@@ -1,12 +1,13 @@
 package com.freshlink.fishdto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record FishResponse(
 		Long id,
         String name,
         String fishType,
-        double pricePerKg,
+        BigDecimal pricePerKg,
         double availableKg,
         boolean inSeason,
         LocalDate seasonStart,
