+++ src/main/java/com/freshlink/fishdto/FishUpdateRequest.java
@@ -1,9 +1,11 @@
 package com.freshlink.fishdto;
 
+import java.math.BigDecimal;
+
 public record FishUpdateRequest(
 			String name,
 	        String fishTypeName,
-	        Double pricePerKg,
+	        BigDecimal pricePerKg,
 	        Double availableKg
 		) {
