+++ src/main/java/com/freshlink/model/OrderItem.java
@@ -1,5 +1,7 @@
 package com.freshlink.model;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -39,5 +41,5 @@ public class OrderItem {
 
 	    
 	    @Column(nullable = false)
-	    private double pricePerKg;
+	    private BigDecimal pricePerKg;
 }
